server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=2.18.2",
    "duckdb>=1.1.3",
    "fastapi>=0.115.6",
    "numpy>=2.2.1",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "ruff>=0.9.2",
    "schedule>=1.2.2",
    "uvicorn>=0.34.0",
]
